Configure project to be able to use pactivate (though this is optional)
We add `pactivate` and `requirements.txt` so that a user can
`source ./pactivate` to create, install and activate the necessary
Python virtual environment as a project-local in one step.

Typically the `Test` script would do this itself, but in order to
allow testing this without other Python configurations, we don't.
Instead, we require the virtualenv to be manually activated before
running `./Test` and print an error message (giving a summary the
three  options and pointing to the README) if `t8dev` is not already
in the path.

Since this sample project is now good for testing both the "no virtual
environment" and the pactivate configurations, and includes a sample
unit test and command-line program build/emulator test, we rename this
to `sample-t8dev-project`.